# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_AhmedHossam777_go-mongo_internal_dto.AuthResponse:
    properties:
      tokens:
        $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.TokenPair'
      user:
        $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.UserResponse'
    type: object
...
  github_com_AhmedHossam777_go-mongo_internal_dto.LoginDto:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  github_com_AhmedHossam777_go-mongo_internal_dto.RefreshTokenInput:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  github_com_AhmedHossam777_go-mongo_internal_dto.RegisterDto:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  github_com_AhmedHossam777_go-mongo_internal_dto.TokenPair:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
  github_com_AhmedHossam777_go-mongo_internal_dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A RESTful API for managing courses, users, and authentication built
    with Go and MongoDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go-MongoDB Course API
  version: "1.0"
paths:
...
  /auth/active-sessions:
    get:
      description: Get all active sessions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of active sessions
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get active sessions
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.AuthResponse'
        "400":
          description: Bad request - invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate refresh token
      parameters:
      - description: Refresh token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - auth
  /auth/refresh-tokens:
    post:
      consumes:
      - application/json
      description: Get new access and refresh tokens using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: New token pair generated
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.TokenPair'
        "401":
          description: Unauthorized - invalid refresh token
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email and password
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.RegisterDto'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/github_com_AhmedHossam777_go-mongo_internal_dto.AuthResponse'
        "400":
          description: Bad request - validation error or user already exists
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"